Minimize code - Remove fragments and page adapters
